refactor: Extract functionality to utils and helpers
