+++ src/index.html
@@ -50,9 +50,11 @@
       </form>
     </header>
 
-    <ul id="historyList"></ul>
+    <ul id="historyList">
+      <li>There are no saved items in history.</li>
+    </ul>
 
-    <button type="button" id="deleteHistoryBtn">Delete history</button>
+    <button type="button" id="deleteHistoryBtn" hidden>Delete history</button>
   </dialog>
 
   <dialog id="settingsDialog" class="popover-dialog popover-dialog--settings">
