+++ src/js/constants.js
@@ -1,0 +1,11 @@
+export const NO_BARCODE_DETECTED = 'No barcode detected';
+
+export const ACCEPTED_MIME_TYPES = [
+  'image/jpg',
+  'image/jpeg',
+  'image/png',
+  'image/apng',
+  'image/gif',
+  'image/webp',
+  'image/avif'
+];
