+++ src/js/helpers/resizeScanFrame.js
@@ -1,0 +1,9 @@
+export function resizeScanFrame(videoEl, scanFrameEl) {
+  if (!videoEl || !scanFrameEl) {
+    return;
+  }
+
+  const rect = videoEl.getBoundingClientRect();
+
+  scanFrameEl.style.cssText = `width: ${rect.width}px; height: ${rect.height}px`;
+}
