+++ src/js/utils/log.js
@@ -1,0 +1,3 @@
+export function log(...args) {
+  process.env.NODE_ENV === 'development' && console.log(...args);
+}
