+++ src/js/utils/vibrate.js
@@ -1,0 +1,11 @@
+export async function vibrate(duration = 0) {
+  if (typeof window.navigator.vibrate !== 'function') {
+    return;
+  }
+
+  try {
+    window.navigator.vibrate(duration);
+  } catch {
+    // Fail silently...
+  }
+}
